// src/components/main/types.tsx
export interface Style {
  direction?: string
  items?: string
  justify?: string
  margin?: string
  width?: string
  height?: string
  overflow?: string
  styleChildren?: string
  wrap?: string
  media?: string
  gap?: string
  mediaChildren?: string
  mHeight?: string
  mWidth?: string
  flex?: string
  padding?: string
}
export interface IFlexProps {
  styleRT: Style
  style?: string
}
export interface ICardMain {
  src: string
  title: string
  content: string
}